+++ src/docs_versions_menu/parse_version.py
@@ -3,6 +3,7 @@
 The ``parse_version`` function provided here is functionally equivalent to
 ``packaging.version.parse`` in ``packaging < 22.0``.
 """
+
 import packaging.version