+++ tests/test_extension/roots/test-rtdtheme/conf.py
@@ -2,7 +2,6 @@ project = 'Basic docs-versions-menu test'
 
 import sphinx_rtd_theme
 
-
 html_theme = "sphinx_rtd_theme"
 html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]